+++ src/tests/integration/server.itest.js
@@ -45,339 +45,6 @@ describe('/api', () => {
   });
 });
 
-describe('/api/proficiencies', () => {
-  it('should list proficiencies', async () => {
-    const res = await request(app).get('/api/proficiencies');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/proficiencies');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/proficiencies?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/proficiencies');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/proficiencies?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/proficiencies/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/proficiencies');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/proficiencies/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/proficiencies/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/races', () => {
-  it('should list races', async () => {
-    const res = await request(app).get('/api/races');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/races');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/races?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/races');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/races?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/races/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/races');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/races/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/races/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-    describe('/api/races/:index/subraces', () => {
-      it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
-        const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/subraces`);
-        expect(res.statusCode).toEqual(200);
-        expect(res.body.results.length).not.toEqual(0);
-      });
-    });
-    describe('/api/races/:index/proficiencies', () => {
-      it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
-        const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/proficiencies`);
-        expect(res.statusCode).toEqual(200);
-        expect(res.body.results.length).not.toEqual(0);
-      });
-    });
-
-    describe('/api/races/:index/traits', () => {
-      it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
-        const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/traits`);
-        expect(res.statusCode).toEqual(200);
-        expect(res.body.results.length).not.toEqual(0);
-      });
-    });
-  });
-});
-
-describe('/api/skills', () => {
-  it('should list skills', async () => {
-    const res = await request(app).get('/api/skills');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/skills');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/skills?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/skills');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/skills?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-  describe('/api/skills/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/skills');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/skills/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/skills/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/spellcasting', () => {
-  it('should list spellcasting', async () => {
-    const res = await request(app).get('/api/spellcasting');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-  describe('/api/spellcasting/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/spellcasting');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/spellcasting/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 30;
-        const showRes = await request(app).get(`/api/spellcasting/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
-describe('/api/spells', () => {
-  it('should list spells', async () => {
-    const res = await request(app).get('/api/spells');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-  });
-
-  it('should hit the cache', async () => {
-    redisClient.del('/api/spells');
-    const clientSet = jest.spyOn(redisClient, 'set');
-    let res = await request(app).get('/api/spells');
-    res = await request(app).get('/api/spells');
-    expect(res.statusCode).toEqual(200);
-    expect(res.body.results.length).not.toEqual(0);
-    expect(clientSet).toHaveBeenCalledTimes(1);
-  });
-
-  describe('with name query', () => {
-    it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/spells');
-      const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/spells?name=${name}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/spells');
-      const name = indexRes.body.results[1].name;
-      const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/spells?name=${queryName}`);
-      expect(res.statusCode).toEqual(200);
-      expect(res.body.results[0].name).toEqual(name);
-    });
-  });
-
-  describe('with level query', () => {
-    it('returns expected objects', async () => {
-      const expectedLevel = 2;
-      const res = await request(app).get(`/api/spells?level=${expectedLevel}`);
-      expect(res.statusCode).toEqual(200);
-
-      const randomIndex = Math.floor(Math.random() * res.body.results.length);
-      const randomResult = res.body.results[randomIndex];
-
-      const indexRes = await request(app).get(`/api/spells/${randomResult.index}`);
-      expect(indexRes.statusCode).toEqual(200);
-      expect(indexRes.body.level).toEqual(expectedLevel);
-    });
-
-    describe('with many provided level', () => {
-      it('returns expected objects', async () => {
-        const expectedLevel1 = 1;
-        const res1 = await request(app).get(`/api/spells?level=${expectedLevel1}`);
-        expect(res1.statusCode).toEqual(200);
-
-        const expectLevel2 = 8;
-        const res2 = await request(app).get(`/api/spells?level=${expectLevel2}`);
-        expect(res2.statusCode).toEqual(200);
-
-        const bothRes = await request(app).get(
-          `/api/spells?level=${expectedLevel1},${expectLevel2}`
-        );
-        expect(bothRes.statusCode).toEqual(200);
-        expect(bothRes.body.count).toEqual(res1.body.count + res2.body.count);
-
-        const randomIndex = Math.floor(Math.random() * bothRes.body.results.length);
-        const randomResult = bothRes.body.results[randomIndex];
-
-        const indexRes = await request(app).get(`/api/spells/${randomResult.index}`);
-        expect(indexRes.statusCode).toEqual(200);
-        expect(
-          indexRes.body.level == expectedLevel1 || indexRes.body.level == expectLevel2
-        ).toBeTruthy();
-      });
-    });
-  });
-
-  describe('with school query', () => {
-    it('returns expected objects', async () => {
-      const expectedSchool = 'Illusion';
-      const res = await request(app).get(`/api/spells?school=${expectedSchool}`);
-      expect(res.statusCode).toEqual(200);
-
-      const randomIndex = Math.floor(Math.random() * res.body.results.length);
-      const randomResult = res.body.results[randomIndex];
-
-      const indexRes = await request(app).get(`/api/spells/${randomResult.index}`);
-      expect(indexRes.statusCode).toEqual(200);
-      expect(indexRes.body.school.name).toEqual(expectedSchool);
-    });
-
-    describe('with many provided schools', () => {
-      it('returns expected objects', async () => {
-        const expectedSchool1 = 'Illusion';
-        const res1 = await request(app).get(`/api/spells?school=${expectedSchool1}`);
-        expect(res1.statusCode).toEqual(200);
-
-        const expectedSchool2 = 'Evocation';
-        const res2 = await request(app).get(`/api/spells?school=${expectedSchool2}`);
-        expect(res2.statusCode).toEqual(200);
-
-        const bothRes = await request(app).get(
-          `/api/spells?school=${expectedSchool1},${expectedSchool2}`
-        );
-        expect(bothRes.statusCode).toEqual(200);
-        expect(bothRes.body.count).toEqual(res1.body.count + res2.body.count);
-
-        const randomIndex = Math.floor(Math.random() * bothRes.body.results.length);
-        const randomResult = bothRes.body.results[randomIndex];
-
-        const indexRes = await request(app).get(`/api/spells/${randomResult.index}`);
-        expect(indexRes.statusCode).toEqual(200);
-        expect(
-          indexRes.body.school.name == expectedSchool1 ||
-            indexRes.body.school.name == expectedSchool2
-        ).toBeTruthy();
-      });
-    });
-
-    it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/spells');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/spells/${index}`);
-      const school = showRes.body.school.name;
-      const querySchool = school.toLowerCase();
-      const res = await request(app).get(`/api/spells?school=${querySchool}`);
-
-      const randomIndex = Math.floor(Math.random() * res.body.results.length);
-      const randomResult = res.body.results[randomIndex];
-
-      const queryRes = await request(app).get(`/api/spells/${randomResult.index}`);
-      expect(queryRes.statusCode).toEqual(200);
-      expect(queryRes.body.school.name).toEqual(school);
-    });
-  });
-
-  describe('/api/spells/:index', () => {
-    it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/spells');
-      const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/spells/${index}`);
-      expect(showRes.statusCode).toEqual(200);
-      expect(showRes.body.index).toEqual(index);
-    });
-    describe('with an invalid index', () => {
-      it('should return one object', async () => {
-        const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/spells/${invalidIndex}`);
-        expect(showRes.statusCode).toEqual(404);
-      });
-    });
-  });
-});
-
 describe('/api/starting-equipment', () => {
   it('should list starting equipment', async () => {
     const res = await request(app).get('/api/starting-equipment');
